# backend/openapi.yaml
openapi: 3.0.3
info:
  title: FinSight API
  description: |
    Personal finance analyzer powered by Plaid and Claude AI.
    Connects to bank accounts via Plaid Link, syncs transactions,
    and generates AI-powered financial insights.

    **Authentication:** Register or login yields a short-lived *access*
    JWT (`Authorization: Bearer …`) plus a longer-lived opaque
    `refresh_token` stored server-side under `SHA-256(raw)` in Postgres.
    Plaid linking, sync, insights, and transaction listing require Bearer auth.
  version: 1.0.0
  contact:
    name: Anthony Vithayathil
    email: anthony.vithayathil1@gmail.com
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://your-railway-url.railway.app
    description: Production, Railway (MVP deployment)
...
paths:
  /api/auth/register:
    post:
      tags: [auth]
      summary: Register a password-backed account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 10
                first_name:
                  type: string
                last_name:
                  type: string
      responses:
        '201':
          description: Account created, returns bearer + rotated refresh payloads
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Duplicate email address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Registration disabled via REGISTRATION_ENABLED=false
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: registration is disabled
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/auth/login:
    post:
      tags: [auth]
      summary: Obtain tokens with password credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenEnvelope'
  /api/auth/refresh:
    post:
      tags: [auth]
      summary: Rotate refresh token and mint a fresh access JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: New opaque refresh token replaces the consumed one atomically in the database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshOnlyEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/auth/logout:
    post:
      tags: [auth]
      summary: Revoke a stored refresh hash (clients should discard bearer tokens locally)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '204':
          description: Successful revoke (still accepts unknown tokens gracefully)
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/link/token:
    get:
      security:
        - BearerAuth: []
      tags: [link]
      summary: Create Plaid link token
      description: |
        Generates a short-lived link_token used to initialize Plaid Link
        on the frontend. The token expires after 30 minutes.
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '200':
          description: Link token created successfully
          content:
            application/json:
              schema:
                type: object
                required: [link_token]
                properties:
                  link_token:
                    type: string
                    example: link-sandbox-abc123
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/link/exchange:
    post:
      security:
        - BearerAuth: []
      tags: [link]
      summary: Exchange public token
      description: |
        Exchanges the temporary public_token received from Plaid Link
        for a permanent access_token. The access_token is encrypted
        and stored server-side, never returned to the client.
        Automatically saves the linked Item to the database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [public_token]
              properties:
                public_token:
                  type: string
                  description: Temporary token from Plaid Link onSuccess callback
                  example: public-sandbox-abc123
                institution_name:
                  type: string
                  description: Name of the linked financial institution
                  example: Chase
      responses:
        '200':
          description: Token exchanged and Item saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Duplicate Plaid item or linked-account limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: institution already linked
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/transactions/sync:
    post:
      security:
        - BearerAuth: []
      tags: [transactions]
      summary: Sync transactions
      description: |
        Syncs transactions for all linked accounts using Plaid's
        /transactions/sync cursor-based endpoint. Handles Plaid
        pagination internally, clients never manage page state.
        After sync, synchronously triggers Claude AI to regenerate
        financial insights. Note: adds 3-5 seconds of latency.
        Production should use async job queue for insight generation.
        All DB operations are atomic.
      responses:
        '200':
          description: Sync completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: sync complete
        '400':
          description: No linked accounts found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: no linked accounts found
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/transactions:
    get:
      security:
        - BearerAuth: []
      tags: [transactions]
      summary: Get all transactions
      description: |
        Returns all synced transactions for the user ordered by
        date descending. Pending transactions are included.
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/items:
    get:
      security:
        - BearerAuth: []
      tags: [accounts]
      summary: List linked institutions
      description: |
        Returns the institutions the authenticated user has linked via Plaid Link.
        Sensitive fields (access token, sync cursor) are never included in this response.
        User identity is resolved from the Bearer token; no user ID in the path.
      responses:
        '200':
          description: List of linked institutions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LinkedItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/accounts:
    get:
      security:
        - BearerAuth: []
      tags: [accounts]
      summary: List accounts
      description: |
        Returns all Plaid accounts for the authenticated user, joined with
        institution name. Balances are refreshed on every transaction sync and
        immediately after linking. User identity is resolved from the Bearer token.
      responses:
        '200':
          description: List of accounts with current balances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/insights:
    get:
      security:
        - BearerAuth: []
      tags: [insights]
      summary: Get financial insights
      description: |
        Returns the most recently generated AI financial insight
        for the user. Insights are generated automatically after
        each transaction sync. Returns a status message if no
        insights have been generated yet.
      responses:
        '200':
          description: Financial insight or no-insight status
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Insight'
                  - type: object
                    properties:
                      status:
                        type: string
                        example: no insights yet
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthTokenEnvelope:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, user]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          format: int64
          description: Access token TTL in seconds from issuance
        user:
          type: object
          required: [id, email]
          properties:
            id:
              type: integer
            email:
              type: string
    RefreshOnlyEnvelope:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          format: int64
    Transaction:
      type: object
      properties:
        ID:
          type: integer
          example: 1
        ItemID:
          type: integer
          example: 1
        UserID:
          type: integer
          example: 1
        PlaidTransactionID:
          type: string
          example: abc123xyz
        Amount:
          type: number
          format: float
          description: Positive = money out, negative = money in
          example: 45.50
        Date:
          type: string
          format: date-time
          example: "2026-04-01T00:00:00Z"
        Name:
          type: string
          description: Raw transaction description
          example: Uber 063015 SF**POOL**
        MerchantName:
          type: string
          description: Cleaned merchant name. Empty string if unavailable.
          example: Uber
        CategoryPrimary:
          type: string
          description: Plaid personal_finance_category primary value
          example: TRANSPORTATION
        CategoryDetailed:
          type: string
          description: Plaid personal_finance_category detailed value
          example: TRANSPORTATION_TAXIS_AND_RIDE_SHARES
        Pending:
          type: boolean
          description: Whether the transaction has settled
          example: false
        AccountID:
          type: integer
          nullable: true
          description: FK to finsight.accounts. Null for transactions synced before accounts were introduced.
          example: 3
    LinkedItem:
      type: object
      required: [id, institution_name, linked_at]
      properties:
        id:
          type: integer
          example: 1
        institution_name:
          type: string
          example: Chase
        linked_at:
          type: string
          format: date-time
          example: "2026-01-15T10:00:00Z"
    Account:
      type: object
      required: [id, item_id, institution_name, name, type]
      properties:
        id:
          type: integer
          example: 3
        item_id:
          type: integer
          example: 1
        institution_name:
          type: string
          example: Chase
        name:
          type: string
          example: Plaid Saving
        official_name:
          type: string
          example: Plaid Gold Standard 0% Interest Saving
        type:
          type: string
          description: "Plaid account type: depository, credit, investment, loan, other"
          example: depository
        subtype:
          type: string
          description: "Plaid account subtype: checking, savings, credit card, etc."
          example: savings
        mask:
          type: string
          description: Last 4 digits of the account number
          example: "1111"
        balance_current:
          type: number
          format: float
          nullable: true
          example: 110.00
        balance_available:
          type: number
          format: float
          nullable: true
          example: 100.00
        iso_currency_code:
          type: string
          example: USD
    CategorySummary:
      type: object
      properties:
        category:
          type: string
          example: FOOD_AND_DRINK
        total_amount:
          type: number
          format: float
          example: 450.00
        count:
          type: integer
          example: 12
    Anomaly:
      type: object
      properties:
        transaction_id:
          type: string
          example: abc123xyz
        name:
          type: string
          example: KFC
        amount:
          type: number
          format: float
          example: 500.00
        reason:
          type: string
          example: A $500 charge at KFC is unusually high for a fast food restaurant
    Insight:
      type: object
      properties:
        ID:
          type: integer
          example: 1
        UserID:
          type: integer
          example: 1
        Summary:
          type: string
          description: 2-3 sentence plain English financial health overview
          example: Your spending is heavily concentrated in food and entertainment.
        TopCategories:
          type: array
          items:
            $ref: '#/components/schemas/CategorySummary'
        Anomalies:
          type: array
          items:
            $ref: '#/components/schemas/Anomaly'
        Recommendations:
          type: array
          items:
            type: string
          example:
            - Review your recurring subscriptions
            - Build a 3-month emergency fund
            - Reduce dining out by 20%
        CreatedAt:
          type: string
          format: date-time
          example: "2026-04-14T22:48:59Z"
        UpdatedAt:
          type: string
          format: date-time
          example: "2026-04-14T22:48:59Z"
    Error:
      type: object
      properties:
        error:
          type: string
          example: something went wrong
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: invalid request body
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: internal server error
    Unauthorized:
      description: Missing or invalid Bearer access token when calling a protected endpoint
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: invalid or expired access token
    Forbidden:
      description: Authenticated but not permitted (e.g. demo account link policy)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: demo account cannot link additional banks
    TooManyRequests:
      description: Rate limit or sync cooldown exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: rate limit exceeded, try again shortly